# swagger.yaml
swagger: "2.0"
info:
  description: "This is a sample server Blog server"
  version: "1.0.0"
  title: "Swagger Node_Blog"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "apiteam@swagger.io"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "localhost:3000"
basePath: "/"
...
schemes:
- "http"
paths:
...
  /posts/like:
    put:
      tags:
      - "post"
      summary: "Update post data"
      description: "Update post by Id"
      operationId: "updatePostById"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Update post"
        required: true
        schema:
          $ref: "#/definitions/AddLike"
      responses:
        default:
          description: "successful operation"
...
definitions:
...
  AddLike:
    type: "object"
    required:
      - "post_id"
      - "user_id"
      - "accessToken"
    properties:
      post_id:
        type: "integer"
        format: "int64"
      user_id:
        type: "integer"
        format: "int64"
      accessToken:
        type: "string"
...